some renaming to fix vscode include path and changes headers to .hpp instead of .h

+++ kernel/src/utils/cstr.cpp
@@ -1,4 +1,4 @@
-#include <strings.h>
+#include <strings.hpp>
 
 // Function to implement strcmp function
 int strcmp(const char *X, const char *Y)
@@ -67,7 +67,7 @@ void memcpy(void *dest, void *src, size_t n){
    char *cdest = (char *)dest;
   
    // Copy contents of src[] to dest[]
-   for (int i = 0; i < n; i++)
+   for (uint64_t i = 0; i < n; i++)
        cdest[i] = csrc[i];
 }
